updated login-related test cases
